ROC-2226: after review
